requirements.txt
langchain
mistralai
prometheus-api-client
jira
PyGithub
python-dotenv
requests
